MS_ControlCenter/requirements.txt
cherrypy
paho-mqtt
requests
python-dotenv